requirements.txt
mysql-connector-python==8.0.28
discord.py==1.7.3